# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PulseFlow API - AI WhatsApp front desk for beauty & wellness businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: serviceId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Conflict - slot unavailable
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/confirm:
    post:
      operationId: confirmBooking
      tags: [bookings]
      summary: Confirm a pending booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingCancelInput"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/complete:
    post:
      operationId: completeBooking
      tags: [bookings]
      summary: Mark booking as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/no-show:
    post:
      operationId: markNoShow
      tags: [bookings]
      summary: Mark booking as no-show
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked no-show
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/reschedule:
    post:
      operationId: rescheduleBooking
      tags: [bookings]
      summary: Reschedule a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingRescheduleInput"
      responses:
        "200":
          description: Rescheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Slot unavailable
  /bookings/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [bookings]
      summary: Get available booking slots for a date and service
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
        - name: serviceId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlotList"

  # ─── Conversations ──────────────────────────────────────────────────────────
...
  /ai/process:
    post:
      operationId: processMessage
      tags: [ai]
      summary: Process inbound customer message through AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMessageInput"
      responses:
        "200":
          description: AI intent result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIIntentResult"
  /ai/actions/{id}/approve:
    post:
      operationId: approveAiAction
      tags: [ai]
      summary: Approve a pending AI action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiApproveInput"
      responses:
        "200":
          description: Action approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIActionLog"
  /ai/actions/{id}/reject:
    post:
      operationId: rejectAiAction
      tags: [ai]
      summary: Reject a pending AI action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Action rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIActionLog"

  # ─── Automation ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, businessId, name, price, durationMinutes, isActive, createdAt]
      properties:
        id: { type: integer }
        businessId: { type: integer }
        name: { type: string }
        category: { type: ["string", "null"] }
        price: { type: number }
        durationMinutes: { type: integer }
        description: { type: ["string", "null"] }
        requiresConsultation: { type: boolean }
        requiresTokenAdvance: { type: boolean }
        repeatReminderDays: { type: ["integer", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Customer:
      type: object
      required: [id, businessId, name, phone, createdAt]
      properties:
        id: { type: integer }
        businessId: { type: integer }
        name: { type: string }
        phone: { type: string }
        notes: { type: ["string", "null"] }
        source: { type: ["string", "null"] }
        lastVisitAt: { type: ["string", "null"] }
        totalVisits: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Booking:
      type: object
      required: [id, businessId, serviceId, bookingDate, startTime, endTime, status, source, createdAt]
      properties:
        id: { type: integer }
        businessId: { type: integer }
        customerId: { type: ["integer", "null"] }
        serviceId: { type: integer }
        bookingDate: { type: string, description: "YYYY-MM-DD" }
        startTime: { type: string, description: "HH:MM" }
        endTime: { type: string, description: "HH:MM" }
        status: { type: string, enum: [pending, confirmed, completed, cancelled, no_show, rescheduled] }
        source: { type: string, enum: [whatsapp, manual, dashboard] }
        notes: { type: ["string", "null"] }
        createdByAI: { type: boolean }
        reminder24hSent: { type: boolean }
        reminder2hSent: { type: boolean }
        reviewRequestSent: { type: boolean }
        repeatReminderScheduledAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        customer:
          $ref: "#/components/schemas/Customer"
        service:
          $ref: "#/components/schemas/Service"
    BookingInput:
      type: object
      required: [serviceId, bookingDate, startTime]
      properties:
        customerId: { type: integer }
        serviceId: { type: integer }
        bookingDate: { type: string }
        startTime: { type: string }
        source: { type: string }
        notes: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
    BookingPatch:
      type: object
      properties:
        notes: { type: string }
        status: { type: string }
    BookingCancelInput:
      type: object
      properties:
        reason: { type: string }
    BookingRescheduleInput:
      type: object
      required: [bookingDate, startTime]
      properties:
        bookingDate: { type: string }
        startTime: { type: string }
        notes: { type: string }
    SlotList:
      type: object
      required: [date, serviceId, slots]
      properties:
        date: { type: string }
        serviceId: { type: integer }
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
    TimeSlot:
      type: object
      required: [startTime, endTime, available]
      properties:
        startTime: { type: string }
        endTime: { type: string }
        available: { type: boolean }

    # ─── Conversation / Message ───────────────────────────────────────────────
...
    AIActionLog:
      type: object
      required: [id, businessId, actionType, status, createdAt]
      properties:
        id: { type: integer }
        businessId: { type: integer }
        customerId: { type: ["integer", "null"] }
        bookingId: { type: ["integer", "null"] }
        conversationId: { type: ["integer", "null"] }
        actionType: { type: string }
        inputSummary: { type: ["string", "null"] }
        outputSummary: { type: ["string", "null"] }
        replyDraft: { type: ["string", "null"] }
        confidenceScore: { type: ["number", "null"] }
        status: { type: string, enum: [pending, approved, rejected, auto_sent] }
        requiresHumanReview: { type: boolean }
        createdAt: { type: string }
        customer:
          $ref: "#/components/schemas/Customer"
    AIIntentResult:
      type: object
      required: [intent, confidence, replyDraft]
      properties:
        intent:
          type: string
          enum: [price_inquiry, availability_inquiry, booking_request, reschedule_request, cancel_request, location_inquiry, faq, unknown]
        confidence: { type: number }
        extractedEntities:
          type: object
          properties:
            serviceName: { type: string }
            requestedDate: { type: string }
            requestedTime: { type: string }
            customerName: { type: string }
        actionSuggestion:
          type: object
          properties:
            shouldReply: { type: boolean }
            shouldCreatePendingBooking: { type: boolean }
            shouldAskFollowUpQuestion: { type: boolean }
            shouldEscalateToOwner: { type: boolean }
        replyDraft: { type: string }
        availableSlots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
        matchedService:
          $ref: "#/components/schemas/Service"
        bookingCreated:
          $ref: "#/components/schemas/Booking"
    AiMessageInput:
      type: object
      required: [message, customerPhone]
      properties:
        message: { type: string }
        customerPhone: { type: string }
        customerName: { type: string }
        conversationId: { type: integer }
    AiApproveInput:
      type: object
      properties:
        editedReply: { type: string }

    # ─── Automation ──────────────────────────────────────────────────────────
...