# lib/api-zod/src/generated/types/aIIntentResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PulseFlow API - AI WhatsApp front desk for beauty & wellness businesses
 * OpenAPI spec version: 0.1.0
 */
import type { AIIntentResultActionSuggestion } from './aIIntentResultActionSuggestion';
import type { AIIntentResultExtractedEntities } from './aIIntentResultExtractedEntities';
import type { AIIntentResultIntent } from './aIIntentResultIntent';
import type { Booking } from './booking';
import type { Service } from './service';
import type { TimeSlot } from './timeSlot';

export interface AIIntentResult {
  intent: AIIntentResultIntent;
  confidence: number;
  extractedEntities?: AIIntentResultExtractedEntities;
  actionSuggestion?: AIIntentResultActionSuggestion;
  replyDraft: string;
  availableSlots?: TimeSlot[];
  matchedService?: Service;
  bookingCreated?: Booking;
}