# lib/api-zod/src/generated/types/aiMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PulseFlow API - AI WhatsApp front desk for beauty & wellness businesses
 * OpenAPI spec version: 0.1.0
 */

export interface AiMessageInput {
  message: string;
  customerPhone: string;
  customerName?: string;
  conversationId?: number;
}